Crate potatosync_notes

Modules

db

embedded_migrations

note

responders

responses

schema

serde_helper

setting

status_response

tag

token

Statics

static_rocket_catch_info_for_token_error
Rocket code generated static catcher info.

static_rocket_route_info_for_add_tag
Rocket code generated static route info.

static_rocket_route_info_for_change_setting
Rocket code generated static route info.

static_rocket_route_info_for_delete_all_notes
Rocket code generated static route info.

static_rocket_route_info_for_delete_note
Rocket code generated static route info.

static_rocket_route_info_for_delete_tag
Rocket code generated static route info.

static_rocket_route_info_for_get_changed_settings
Rocket code generated static route info.

static_rocket_route_info_for_get_deleted_notes
Rocket code generated static route info.

static_rocket_route_info_for_get_deleted_tags
Rocket code generated static route info.

static_rocket_route_info_for_get_notes
Rocket code generated static route info.

static_rocket_route_info_for_get_setting
Rocket code generated static route info.

static_rocket_route_info_for_get_tags
Rocket code generated static route info.

static_rocket_route_info_for_patch_note
Rocket code generated static route info.

static_rocket_route_info_for_patch_tag
Rocket code generated static route info.

static_rocket_route_info_for_ping
Rocket code generated static route info.

static_rocket_route_info_for_save_note
Rocket code generated static route info.

static_rocket_route_info_for_secure_ping
Rocket code generated static route info.

static_rocket_route_info_for_update_note
Rocket code generated static route info.

Functions

add_tag

change_setting
Route for changing setting

delete_all_notes
Route for deleting all notes of an user

delete_note
Route for deleting single note identified by id

delete_tag

get_changed_settings
Route for getting changed key-value pairs of settings

get_deleted_notes
Route for getting a list of deleted notes based on a provided list of id's

get_deleted_tags
Route for getting a list of deleted notes based on a provided list of id's

get_notes
Route for getting all the notes which are updated after provided timestamp

get_setting
Route for getting setting

get_tags

main

patch_note

patch_tag

ping
Route for checking connectivity

rocket_catch_fn_token_error
Rocket code generated wrapping catch function.

rocket_route_fn_add_tag
Rocket code generated wrapping route function.

rocket_route_fn_change_setting
Rocket code generated wrapping route function.

rocket_route_fn_delete_all_notes
Rocket code generated wrapping route function.

rocket_route_fn_delete_note
Rocket code generated wrapping route function.

rocket_route_fn_delete_tag
Rocket code generated wrapping route function.

rocket_route_fn_get_changed_settings
Rocket code generated wrapping route function.

rocket_route_fn_get_deleted_notes
Rocket code generated wrapping route function.

rocket_route_fn_get_deleted_tags
Rocket code generated wrapping route function.

rocket_route_fn_get_notes
Rocket code generated wrapping route function.

rocket_route_fn_get_setting
Rocket code generated wrapping route function.

rocket_route_fn_get_tags
Rocket code generated wrapping route function.

rocket_route_fn_patch_note
Rocket code generated wrapping route function.

rocket_route_fn_patch_tag
Rocket code generated wrapping route function.

rocket_route_fn_ping
Rocket code generated wrapping route function.

rocket_route_fn_save_note
Rocket code generated wrapping route function.

rocket_route_fn_secure_ping
Rocket code generated wrapping route function.

rocket_route_fn_update_note
Rocket code generated wrapping route function.

save_note
Route for saving note

secure_ping
Route for checking if the user is logged in

token_error
Route used for catching 401 errors e.g. invalid access token

update_note
Route for updating note