Module potatosync_notes::api::endpoints::notes

Constants

INTERNAL_DB_ERROR

NOTES_DELETE_SUCCESS

NOTE_ADD_SUCCESS

NOTE_ALREADY_EXISTS

NOTE_DELETE_SUCCESS

NOTE_LIST_SUCCESS

NOTE_MISSING_LAST_MODIFY

NOTE_NOT_EXISTS

NOTE_NOT_FOUND

NOTE_PATCH_SUCCESS

NOTE_UPDATE_SUCCESS

Statics

static_rocket_route_info_for_create
Rocket code generated static route info.

static_rocket_route_info_for_delete
Rocket code generated static route info.

static_rocket_route_info_for_delete_all
Rocket code generated static route info.

static_rocket_route_info_for_get_deleted
Rocket code generated static route info.

static_rocket_route_info_for_get_updated
Rocket code generated static route info.

static_rocket_route_info_for_patch
Rocket code generated static route info.

static_rocket_route_info_for_update
Rocket code generated static route info.

Functions

create
Route for saving note

delete
Route for deleting single note identified by id

delete_all
Route for deleting all notes of an user

fuel

get_deleted
Route for getting a list of deleted notes based on a provided list of id’s

get_updated
Route for getting all the notes which are updated after provided timestamp

patch

rocket_route_fn_create
Rocket code generated wrapping route function.

rocket_route_fn_delete
Rocket code generated wrapping route function.

rocket_route_fn_delete_all
Rocket code generated wrapping route function.

rocket_route_fn_get_deleted
Rocket code generated wrapping route function.

rocket_route_fn_get_updated
Rocket code generated wrapping route function.

rocket_route_fn_patch
Rocket code generated wrapping route function.

rocket_route_fn_update
Rocket code generated wrapping route function.

update
Route for updating note