Module potatosync_notes::note::controller

Controller to handle the notes

Structs

DeletedResponse
Struct used when server needs to return list of notes along with status of request

NoteResponse
Struct used when server needs to return list of notes along with status of request

Functions

add
Adds note and if they already exist, it will do nothing

delete
Delete single note in DB

delete_all
Delete all notes of an user

get_deleted_by_account

get_notes_by_account
Get list of notes updated after provided timestamp

patch
Patches note in database. It will only change the fields provided

update
Updates note, it will create a new note if it doesnt exist