Module potatosync_notes::schemas::notes

Structs

DeletedResponse
Struct used when server needs to return list of notes along with status of request

NoteId
Struct used when client needs to specify certain note_id

NoteLastUpdated
Struct used when client needs to specify last_updated timestamp

NoteResponse

PatchingNote

SavingNote
Note as provided by the client when saving. Note the missing account_id since the client doesnt know the id.